+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - Only updating relevant displays presenters halves routing feedback time
  - CUE Pages will now properly be hidden in rooms without a scheduler
  - Complete overhaul of settings pages
+ - Performance improvement for routing in single display rooms
 
 ## [9.1.0] - 2019-08-27
 ### Added